Fixing a Problem with the Delete Button

+++ holdings-tab.go
@@ -13,9 +13,14 @@ import (
 )
 
 func (app *Config) holdingsTab() *fyne.Container {
+	app.Holdings = app.getHoldingSlice()
 	app.HoldingsTable = app.getHoldingsTable()
+
 	holdingsContainer := container.NewBorder(
-		nil, nil, nil, nil,
+		nil,
+		nil,
+		nil,
+		nil,
 		container.NewAdaptiveGrid(1, app.HoldingsTable),
 	)
 
@@ -23,42 +28,44 @@ func (app *Config) holdingsTab() *fyne.Container {
 }
 
 func (app *Config) getHoldingsTable() *widget.Table {
-	data := app.getHoldingSlice()
-	app.Holdings = data
-
 	t := widget.NewTable(
 		func() (int, int) {
-			return len(data), len(data[0]) // rows, columns
+			return len(app.Holdings), len(app.Holdings[0])
 		},
 		func() fyne.CanvasObject {
 			ctr := container.NewVBox(widget.NewLabel(""))
 			return ctr
 		},
 		func(i widget.TableCellID, o fyne.CanvasObject) {
-			if i.Col == (len(data[0])-1) && i.Row != 0 {
+			if i.Col == (len(app.Holdings[0])-1) && i.Row != 0 {
 				// last cell - put in a button
 				w := widget.NewButtonWithIcon("Delete", theme.DeleteIcon(), func() {
 					dialog.ShowConfirm("Delete?", "Are you sure you want to delete this holding?", func(deleted bool) {
-						id, _ := strconv.Atoi(data[i.Row][0].(string))
-						err := app.DB.DeleteHolding(int64(id))
-						if err != nil {
-							app.ErrorLog.Println(err)
+						if deleted {
+							id, _ := strconv.Atoi(app.Holdings[i.Row][0].(string))
+							err := app.DB.DeleteHolding(int64(id))
+							if err != nil {
+								app.ErrorLog.Println(err)
+							}
 						}
 						// refresh the holdings table
 						app.refreshHoldingsTable()
 					}, app.MainWindow)
 				})
 				w.Importance = widget.HighImportance
-				o.(*fyne.Container).Objects = []fyne.CanvasObject{w}
-			} else {
-				// just put in textual information
+
 				o.(*fyne.Container).Objects = []fyne.CanvasObject{
-					widget.NewLabel(data[i.Row][i.Col].(string)),
+					w,
+				}
+			} else {
+				// we're just putting in textual information
+				o.(*fyne.Container).Objects = []fyne.CanvasObject{
+					widget.NewLabel(app.Holdings[i.Row][i.Col].(string)),
 				}
 			}
 		})
 
-	colWidths := []float32{50, 200, 200, 200, 110}
+	colWidths := []float32{50, 200, 200, 200, 100}
 	for i := 0; i < len(colWidths); i++ {
 		t.SetColumnWidth(i, colWidths[i])
 	}
